backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
gunicorn==21.2.0
azure-keyvault-secrets==4.7.0
azure-identity==1.15.0
sqlalchemy==2.0.23
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
python-dateutil==2.8.2
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2

